Added PodRemoteMixin
