+++ app/controllers/container_group_controller.rb
@@ -1,7 +1,7 @@
 class ContainerGroupController < ApplicationController
   include Mixins::ContainersCommonMixin
   include Mixins::BreadcrumbsMixin
-  include PodRemote
+  include Mixins::PodRemoteMixin
 
   before_action :check_privileges
   before_action :get_session_data
